feat: provide support for Holtek HT16K33 (#14)
Provide support for the Holtek HT16K33 (LED Controller Driver with
keyscan).

This chip is used by various 7-segment, 14-segment and LED matrix
backpacks made by Adafruit. A low-level and a higher-level interface is
provided. The higher-level interface makes usage more convenient (e.g.
user does not need to worry about the correct ordering of operations).

- [x] provide an example for Espressif ESP32-C6
- [x] provide an example for Raspberry Pi Pico
- [x] provide support for std environments (required for `cargo test`)
- [x] provide a virtual I²C bus device
- [x] use virtual I²C bus device to provide unit tests
- [x] test all hardware functions for HT16K33
- [x] test all semantic functions for HT16K33

It is unclear how to use the "13×3 matrix key scanning" functionality.
This feature is not implemented.